Move rally transforms into a folder
